refactor: cleanup verification logic
